Updated for pygit2-0.19, Refactored decl. models
* Updated everything to work with pygit2-0.19
* Refactored how models are declared outside of workspace.
* Tests passing

A model that subclasses GitModel must now always define a __workspace__
attribute. Otherwise, it must subclass an existing workspace model or
GitMmodelBase. GitModelBase is not metaclassed, but once you load a GitModelBase
model into a workspace it is rebuilt into a metaclassed version of that model.
This helps with various issues around the previous method of creating registered
models as subclasses of the original.